labgrid/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "labgrid-usb-device"
version = "0.1.0"
description = "Labgrid integration for usb-device managed hardware"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "labgrid>=24.0",
    "attrs>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
